Fixed infinite window resizing issue.
Was unable to properly run the menu because main window would constantly resize every frame by exactly 16 pixels. Discovered that for some reason the setPreferredSize method at line 101 was adding exactly 16 pixels to rectangle1.width every time it was called, resulting in rectangle1 never being equal to rectangle2. Added some padding for comparing the width and height of the two rectangles seemed to resolve the issue for now and should make the menu less laggy. Would like to explore what is causing the discrepancy in the setPreferredSize method later.